Make standalone and bundled spatGRIS better behaved and more differenciated
This irons out some of the kinks of SpeakerView when it is started independently.

When SpeakerView is started from SpatGRIS, this

    Disables the OSC server when SpeakerView is started from SpatGRIS
    Disables the networking options when SpeakerView is started from SpatGRIS

This makes it so users of SpatGRIS do not have to worry about external OSC messages messing up their SpeakerView and makes one less icon onscreen.

When SpeakerView is started independently, this makes it

    Ignore SpatGRIS's requests to shutdown
    Ignore SpatGRIS's requests to do windowing focus/order manipulations

This also changes the default way godot launches SpeakerView so that the standalone version is the default.