Merge SV1-prepare-release into dev (#28)
This brings the last changes into dev for the next release.

* Add logos to README.md

* update logo

* Adds build pipelines to speakerview  (#4)

* tries to add pipelines

* add export presets

* good name for Linux template

* Make pipelines run on ubuntu 24.04 (#6)

* build on ubuntu latest

* runs on 24.04 instead of latest

* Add a free camera mode in addition to the default orbit camera (#7)

Adds a free camera mode and a toggle button to switch between the two cameras. The default camera is still the orbit camera.

* Add a documentation panel which lists keybindings (#8)

Adds a documentation panel which documents global keybindings and keybindings for the different camera modes.

This also fixes a bug where the keyboard shortcuts would have no effect until `update_app_data` would be called for an unrelated reason.

* Add OSC support and makes SpeakerView launchable without spatGRIS (#9)

This adds support for externally controlling SpeakerView with OSC messages.

    Adds support for source and speakers control by OSC
    Adds support for controlling certain software features from OSC
    Adds a visual "test" of every implemented OSC feature in the new tests folder
    Changes the pop-up text at startup to explain that SpeakerView can be used standalone but that its not required for a simple usage of SpatGRIS

Of note :

    This fixes a source related bug I had introduced with the free camera (woops)
    This fixes a mislabeled keybinding in the help panel (woops)
    This moves a lot of display logic to the _process functions of speakers and sources. This logic was previously executed on reception of network packets and a comment seemed to talk about performance issue with _process but I was able to display (3) thousands of sources and destinations on my moderately powerful laptop with a decent framerate with the code as is.

* makes it possible to select speakers with the free camera (#10)

* Auto-scale speakers based on their collisions (#11)

This adds speaker autoscaling based on whether or not they are intersecting with neighboring speakers.

Of note:

    This was created so that tight speaker polygons are automatically displayed in a more elegant fashion.
    This does not scale the speakers up again when they do not collide anymore. It shouldn't be too much of a problem since speaker setups are mostly static once edited.

* Fix a Speakerview freeze when a speaker is centered (#12)

This creates a `Utils.safe_look_at` which does the necessary checks to avoir throwing errors every frame in specific circumstances.

When SpeakerView is started by spatGRIS, this error flood used to quickly freezes SpeakerView.

* Make speakers that are part of a group to be oriented towards the center of that group (#13)

Uses the informations added by https://github.com/GRIS-UdeM/SpatGRIS/pull/489 to improve the display of speaker groups.

* Adds settings field to change upd output and input ports/addresses

Also includes small changes that results in less warnings in the editor.

* Make standalone and bundled spatGRIS better behaved and more differenciated



This irons out some of the kinks of SpeakerView when it is started independently.

When SpeakerView is started from SpatGRIS, this

    Disables the OSC server when SpeakerView is started from SpatGRIS
    Disables the networking options when SpeakerView is started from SpatGRIS

This makes it so users of SpatGRIS do not have to worry about external OSC messages messing up their SpeakerView and makes one less icon onscreen.

When SpeakerView is started independently, this makes it

    Ignore SpatGRIS's requests to shutdown
    Ignore SpatGRIS's requests to do windowing focus/order manipulations

This also changes the default way godot launches SpeakerView so that the standalone version is the default.

* Adds a 'velocity' deadzone to mouse capture and adjust drag physics

Addresses https://github.com/GRIS-UdeM/SpeakerView/issues/16

* Make speakerview wait for 1 spatGRIS app state packet before sending its own (#18)

This fixes a pre-existing race condition that was made more common by the change that makes spatGRIS spam less udp messages.

* Fix Flickering Bug

SpeakerView used to send its state every time it received anything from spatGRIS.
This caused a bug where settings would rapidly turn on and off.

* Adjust some free camera mouse behaviours  (#20)

* invert left and right click and dont capture cursor in free cam

* remove mouse velocity when not clicking

* sensitivity changes

* Adds arm64 CI build (#22)

* Add fulldome camera mode

This adds a fulldome camera mode that is toggled with alt+d. This mode is documented in the help panel.

* Add MacOS codesigning (#24)

* Deactivate embed pck for linux (#26)

I had accidentally pushed a commit with embed pck activated in the Linux export template. This broke the SpatGRIS integration because SpatGRIS looks for a pck file on linux and windows and refuses to launch spatgris if it doesn't exist.

* Upgrade to godot4.5 (#27)

This upgrades SpeakerView to Godot4.5.

It also enables the new shader baker which can lower startup times and it changes the CI image provider to something up to date (which resulted in a partial rewrite of the actions file.)

* Updated Readme

* Updated SpeakerView to version 1.0.0

* SV new logo

* Regenarated UIDs (Godot Project -> Tools -> Upgrade UIDs...)

* Updated icon.ico with new logo

* Rebuilt UIDs on MacOS...

* Icon format is .png for MacOS and .ico for Windows

* New log again

* Fix wrong keybinding name in help panel

Its source activity not speaker activity

---------

Co-authored-by: Vincent Berthiaume <vberthiaume@gmail.com>
Co-authored-by: Guillaume Riou <117857232+guillaumeriousat@users.noreply.github.com>